Changed the recent method of the user class to function more efficiently
I changed the recent method of the user class so that it seeks out the
uid and from that makes a for loop to get the post ids of any posts
that come before it. Takes the 100 most recent posts.

I also fixed the for loop that selects the posts. When looping with the
xrange function it includes the first number and goes until but
excludes the second number. Before it wasn't posting the most recent
post. So it should be working properly now. (If this is wrong for some
reason, let me know, it worked fine on my retwis-py).