More hashtag support, users page
Going to rickshould.servebeer.com/hashtag/<inserttag> will will lead to
a page that contains any posts associated with a specific hashtag in
chronological order.

rickshould.servebeer.com/users will return a list of all users in the
database in alphabetical order.

Anyway, lots of new things to play with an improve upon!